+++ tests/unit/test_calculate_job_count_for_tiny_values.py
@@ -1,0 +1,39 @@
+import unittest
+import warnings
+
+from pyspark.sql import SparkSession
+from pyspark.sql.types import StructType, StructField, StringType, IntegerType, DoubleType
+
+from utils.prepare_locations_utils.job_calculator.calculate_job_count_for_tiny_values import \
+    calculate_jobcount_handle_tiny_values
+
+
+class TestJobCountTinyValues(unittest.TestCase):
+    calculate_jobs_schema = StructType(
+        [
+            StructField("locationid", StringType(), False),
+            StructField("total_staff", IntegerType(), True),
+            StructField("worker_record_count", IntegerType(), True),
+            StructField("number_of_beds", IntegerType(), True),
+            StructField("job_count", DoubleType(), True),
+        ]
+    )
+
+    def setUp(self):
+        self.spark = SparkSession.builder.appName(
+            "test_ahndle_tiny_values"
+        ).getOrCreate()
+
+        warnings.simplefilter("ignore", ResourceWarning)
+
+    def test_calculate_jobcount_handle_tiny_values(self):
+        rows = [
+            ("1-000000008", 2, 53, 26, None),
+        ]
+        df = self.spark.createDataFrame(data=rows, schema=self.calculate_jobs_schema)
+
+        df = calculate_jobcount_handle_tiny_values(df)
+        self.assertEqual(df.count(), 1)
+
+        df = df.collect()
+        self.assertEqual(df[0]["job_count"], 53)
