+++ tests/unit/test_prepare_locations.py
@@ -22,12 +22,7 @@ from tests.test_file_generator import (
     generate_ons_denormalised_data,
     generate_pir_file,
 )
-from utils.prepare_locations_utils.job_calculator import (
-    calculate_jobcount_estimate_from_beds,
-)
-from utils.prepare_locations_utils.job_calculator.calculate_job_count_for_tiny_values import (
-    calculate_jobcount_handle_tiny_values,
-)
+
 from utils.prepare_locations_utils.job_calculator.calculate_jobcount_abs_difference_within_range import (
     calculate_jobcount_abs_difference_within_range,
 )
@@ -440,62 +435,7 @@ class PrepareLocationsTests(unittest.TestCase):
         self.assertEqual(df[0]["job_count"], 11)
         self.assertEqual(df[1]["job_count"], 104.5)
 
-    def test_calculate_jobcount_handle_tiny_values(self):
-        rows = [
-            ("1-000000008", 2, 53, 26, None),
-        ]
-        df = self.spark.createDataFrame(data=rows, schema=self.calculate_jobs_schema)
 
-        df = calculate_jobcount_handle_tiny_values(df)
-        self.assertEqual(df.count(), 1)
-
-        df = df.collect()
-        self.assertEqual(df[0]["job_count"], 53)
-
-    def test_calculate_jobcount(self):
-        columns = ["locationid", "worker_record_count", "total_staff", "number_of_beds"]
-        rows = [
-            ("1-000000001", None, 0, 0),  # Both 0: Return 0
-            # Both 500: Return 500
-            ("1-000000002", 500, 500, 490),
-            # Only know worker_record_count: Return worker_record_count (100)
-            ("1-000000003", 100, None, 10),
-            # Only know total_staff: Return totalstaf (10)
-            ("1-000000004", None, 10, 12),
-            # None of the rules apply: Return None
-            ("1-000000005", 25, 75, 40),
-            # None of the rules apply: Return None
-            ("1-000000006", 30, 60, 40),
-            # None of the rules apply: Return None
-            ("1-000000007", 600, 900, 150),
-            # Absolute difference is within 10%: Return Average
-            ("1-000000008", 10, 12, None),
-            # Either total_staff or worker_record_count < 3: return max
-            ("1-000000009", 1, 23, None),
-            # Utilise bedcount estimate - Average
-            ("1-000000010", 90, 102, 85),
-            # Utilise bedcount estimate - Wkrrecs
-            ("1-000000011", 90, 102, 95),
-            # Utilise bedcount estimate - Totalstaff
-            ("1-000000012", 90, 102, 80),
-        ]
-        df = self.spark.createDataFrame(rows, columns)
-
-        jobcount_df = prepare_locations.calculate_jobcount(df)
-        jobcount_df_list = jobcount_df.collect()
-
-        self.assertEqual(jobcount_df_list[0]["job_count"], 0.0)
-        self.assertEqual(jobcount_df_list[1]["job_count"], 500.0)
-        self.assertEqual(jobcount_df_list[2]["job_count"], 100.0)
-        self.assertEqual(jobcount_df_list[3]["job_count"], 10.0)
-        self.assertEqual(jobcount_df_list[4]["job_count"], None)
-        self.assertEqual(jobcount_df_list[5]["job_count"], None)
-        self.assertEqual(jobcount_df_list[6]["job_count"], None)
-        self.assertEqual(jobcount_df_list[7]["job_count"], 11.0)
-        self.assertEqual(jobcount_df_list[8]["job_count"], 23.0)
-        self.assertEqual(jobcount_df_list[9]["job_count"], 96.0)
-        self.assertEqual(jobcount_df_list[10]["job_count"], 102.0)
-        self.assertEqual(jobcount_df_list[11]["job_count"], 90.0)
 
     def test_get_cqc_location_df_standardises_yorkshire_and_the_humber_region(self):
         cqc_locations = prepare_locations.get_cqc_location_df(
